FullStack/py-app/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
fastapi==0.115.6
graphql-core==3.2.5
idna==3.10
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.41.3
strawberry-graphql==0.258.0
typing_extensions==4.12.2
